+++ core/src/texture.cpp
@@ -3,8 +3,8 @@
 #include <utils/log.hpp>
 
 #include <SDL3_image/SDL_image.h>
-#include <nlohmann/json.hpp>
 #include <fstream>
+#include <nlohmann/json.hpp>
 
 namespace core {
 
@@ -30,7 +30,7 @@ Texture::Texture(const std::string& asset_name)
     : _asset_name{asset_name},
       _current_row{},
       _current_col{},
-      _animation_speed{1}{
+      _animation_speed{1} {
 
     static auto all_assets = assets();
     auto asset = all_assets[asset_name];
@@ -39,7 +39,8 @@ Texture::Texture(const std::string& asset_name)
     _frame_width = asset["frame_size"]["width"];
     _frame_height = asset["frame_size"]["height"];
 
-    _sdl_texture = IMG_LoadTexture(Window::instance().sdl_renderer(), asset_path.data());
+    _sdl_texture =
+        IMG_LoadTexture(Window::instance().sdl_renderer(), asset_path.data());
     float width{}, height{};
     SDL_GetTextureSize(_sdl_texture, &width, &height);
 
@@ -49,8 +50,8 @@ Texture::Texture(const std::string& asset_name)
 }
 
 void Texture::update() {
-    size_t ticks = static_cast<size_t>(SDL_GetTicks() / 50);
-    size_t frameNumber = _animation_speed * ticks;
+    int ticks = static_cast<int>(SDL_GetTicks() / 50);
+    int frameNumber = _animation_speed * ticks;
     _current_col = frameNumber % _col_count;
     _current_row = frameNumber / _col_count % _row_count;
 }
@@ -60,16 +61,17 @@ void Texture::render(Rect position) {
     SDL_FRect source{_frame_width * static_cast<float>(_current_col),
                      _frame_height * static_cast<float>(_current_row),
                      _frame_width, _frame_height};
-    SDL_FRect sdl_destination{position[0], position[1], position[2],
-                              position[3]};
+    SDL_FRect sdl_destination{
+        static_cast<float>(position[0]), static_cast<float>(position[1]),
+        static_cast<float>(position[2]), static_cast<float>(position[3])};
 
     decltype(auto) renderer = Window::instance().sdl_renderer();
     SDL_RenderTexture(renderer, _sdl_texture, &source, &sdl_destination);
 }
 
-// size_t TextureManager::_instance_count = 0;
+// int TextureManager::_instance_count = 0;
 
-// TextureManager::TextureManager() 
+// TextureManager::TextureManager()
 //     : _textures{} {
 //     if (_instance_count++ != 0)
 //         throw std::runtime_error("only one instance of Texture Manager can be instanciated");
